feat: improve log format
