--- apps/cloud/src/tracing.rs
@@ -1,38 +1,0 @@
-use super::*;
-use axum::{body::Body, http::Request, Router};
-use nanoid::nanoid;
-use tower_http::{
-    request_id::{MakeRequestId, PropagateRequestIdLayer, RequestId, SetRequestIdLayer},
-    trace::TraceLayer,
-};
-
-#[derive(Clone, Copy)]
-pub struct MakeRequestUuid;
-
-impl MakeRequestId for MakeRequestUuid {
-    fn make_request_id<B>(&mut self, _request: &Request<B>) -> Option<RequestId> {
-        let request_id = nanoid!().parse().unwrap();
-        Some(RequestId::new(request_id))
-    }
-}
-
-pub fn make_tracing_layer(router: Router) -> Router {
-    router
-        .layer(
-            TraceLayer::new_for_http().make_span_with(|request: &Request<Body>| {
-                let request_id = request
-                    .extensions()
-                    .get::<RequestId>()
-                    .and_then(|id| id.header_value().to_str().ok())
-                    .unwrap_or_default();
-                info_span!(
-                    "HTTP",
-                    http.method = %request.method(),
-                    http.url = %request.uri(),
-                    request_id = %request_id,
-                )
-            }),
-        )
-        .layer(PropagateRequestIdLayer::x_request_id())
-        .layer(SetRequestIdLayer::x_request_id(MakeRequestUuid))
-}
