+++ libs/jwst-logger/src/lib.rs
@@ -5,7 +5,7 @@ mod logger;
 pub use logger::init_logger;
 pub use tracing::{
     debug, debug_span, error, error_span, info, info_span, instrument, log::LevelFilter, trace,
-    trace_span, warn, warn_span,
+    trace_span, warn, warn_span, Level,
 };
 
 pub use tracing;
